requirements.txt
groq
requests
beautifulsoup4
pydub
moviepy
pytrends
yt-dlp
edge-tts
python-dotenv
audioop-lts
gspread
oauth2client
